fix: update the 'GetOverallRating' method in Vendor
Update the method to account for the change
from int to enum in the 'Review' entity

+++ BiteAlert/Modules/VendorModule/Vendor.cs
@@ -17,6 +17,6 @@ public class Vendor : ApplicationUser
     {
         return Reviews.Count == 0 ?
             0 :
-            Reviews.Average(r => r.Rating);
+            Reviews.Average(r => (int)r.Rating);
     }
 }
